# openapi.yaml
openapi: 3.0.2
info:
  title: Voluble API
  version: '1.0'
  description: >
    This is the OpenAPI specification for the Voluble messaging service API.

    Current API status:![OpenAPI Validity
    Status](http://validator.swagger.io/validator?url=https://raw.githubusercontent.com/calmcl1/voluble/master/openapi.json)
    
    Note: all responses are `JSEND`-compatible, which means that the general success or failure of a request can be determined from the `status` field in the HTTP response. The API data returned can be found in the `data` field in the response.
  contact:
    email: mail@callum-mclean.co.uk
    name: Cal McLean
servers:
  - url: 'https://voluble-poc.herokuapp.com/v1'
...
components:
  responses:
    '400':
      description: Data supplied is invalid
    '404':
      description: No object found with the specified ID
    '410':
      description: Object with specified ID has already been deleted
    '500':
      description: Internal server error
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: string
          nullable: false
          readOnly: true
          description: The UUID of the Organization within Voluble
          example: c616d114-4f68-4dba-b976-cd5ae0aa359d
        name:
          type: string
          description: The human-friendly name of the Organization
          nullable: false
          example: MyCorp
        phone_number:
          type: string
          description: >-
            The E164-formatted phone number where all SMSs from Voluble on
            behalf of the Organization will originate
          nullable: false
          example: '+442071838750'
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'A JWT bearer token generated by Auth0, representing the User.'
paths:
...
  /organizations:
    get:
      tags:
        - Organizations
      summary: Returns a list of all of the Organizations available
      description: Returns a list of all of the Organizations available
      security:
        - jwt: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Organization'
                maxItems: 100
        '500':
          $ref: '#/components/responses/500'
    post:
      tags:
        - Organizations
      summary: Creates a new Organizations
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  '/organizations/{organization_id}':
    parameters:
      - name: organization_id
        description: The UUID of the Organization within Voluble to manipulate
        required: true
        in: path
        schema:
          type: string
    get:
      tags:
        - Organizations
      security:
        - jwt: []
      summary: Retrieves the Organization with the specified ID
      responses:
        '200':
          description: Organization retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Organizations
      security:
        - jwt: []
      summary: Modifies the Organization with the specified ID
      requestBody:
        required: true
        description: The new data to replace the specified parameters with
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
      responses:
        '200':
          description: Organization modified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Organizations
      security:
        - jwt: []
      summary: Removes the Organization with the specified ID
      responses:
        '200':
          description: Organization deleted successfully
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '410':
          $ref: '#/components/responses/410'
        '500':
          $ref: '#/components/responses/500'